r39 == r17_beta19(2011-12-04) 231_r39_kene.zip (斧のみ)
進行不能バグを解消したような気がするので更新します。
基本的に前回と殆ど変わっていません。

† 進行不能バグを解消したと思う。
† 自機を追いかけるようなザコは本家にないので廃止。
† ザコの乱数も本家にないのでなるべく廃止する方向で。
† ザコ領域とか使ってないのに無駄に増えてる(256->512)部分は前回より遅くなってるかも。
† ゲーム中右サイドの数字のうち「ハイスコア」、「スコア」、「グレイズ」の3つだけGu化した。(残りは数字含めまだSDL)
† ライブラリとか地味に修正。
† スペカとか1枚も増えてない。
† 表示関係のバグはまだ沢山ある。
† 他にもあったような気がするが思い出せない。
† その他。

+++ trunk/src/game_core/include/000_support.h
@@ -329,9 +329,9 @@ R8G8B8A8フォーマットで色を設定する事になりますが、PSPはリトルエンディアンのCPUを
 #define t8_floor(x) 			((x)>>(3))
 
 /* 秒単位の時間で */
-#define byou60(x)	(x*60)
+#define byou60(x)	((int)(x*60))
 /* 約秒単位の時間で */
-#define byou64(x)	(x*64)
+#define byou64(x)	((int)(x*64))
 
 /* aaa%の確率で */
 #define rand_percent(aaa) ((unsigned char)(ra_nd())<=(unsigned char)( (aaa*256)/100 ) ) /* aaa%の確率で */
